refactor: move index (root) into src dir; rename old index to client
